refactor: remove empty components
