+++ src/routes/about-us/components/AboutCore.svelte
